+++ CHANGELOG.md
@@ -5,12 +5,13 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
-## Version 1.11.1 - 2026-xx-xx
+## Version 1.11.1 - 2026-08-04
 
 ### Fixed
 
 - Migrate jest to vitest
 - Fix `InsertResult` for CURD
+- Dependencies
 
 ## Version 1.11.0 - 2026-07-03